+++ src/pages/members.tsx
@@ -8,23 +8,9 @@ import Seo from "../components/Seo"
 import PageHeader from "../components/PageHeader"
 import ContactUs from "../components/ContactUs"
 import Logos, { LogoType } from "../components/Logos"
+import { scrollToSection } from "../util/scrollToView"
 
 const Members = () => {
-  const scrollToSection = (
-    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-    sectionId: string,
-  ) => {
-    event.preventDefault()
-    const section = document.getElementById(sectionId)
-    if (section) {
-      const navbarHeight = 98
-      window.scrollTo({
-        top: section.offsetTop - navbarHeight,
-        behavior: "instant",
-      })
-    }
-  }
-
   return (
     <Layout>
       <PageHeader
